JNG-3176 Move to KarafProvider (Java 11 compiliant) ITest variant (#2)
