+++ osgi-email-itest/pom.xml
@@ -13,7 +13,6 @@
     <description>Email services karaf integration tests</description>
 
     <dependencies>
-
         <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-context-support</artifactId>
@@ -119,24 +118,6 @@
         </dependency>
 
         <dependency>
-            <groupId>hu.blackbelt.karaf.features</groupId>
-            <artifactId>google-features</artifactId>
-            <classifier>features</classifier>
-            <type>xml</type>
-            <scope>test</scope>
-            <version>${google-karaf-features-version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>hu.blackbelt.karaf.features</groupId>
-            <artifactId>subethamail-features</artifactId>
-            <classifier>features</classifier>
-            <type>xml</type>
-            <scope>test</scope>
-            <version>${subethamail-karaf-features-version}</version>
-        </dependency>
-
-        <dependency>
             <groupId>hu.blackbelt</groupId>
             <artifactId>osgi-email-karaf-features</artifactId>
             <classifier>features</classifier>
@@ -191,12 +172,15 @@
 
                 <configuration>
                     <forkedProcessExitTimeoutInSeconds>1</forkedProcessExitTimeoutInSeconds>
+                    <systemPropertyVariables>
+                        <karafVersion>${karaf-version}</karafVersion>
+                    </systemPropertyVariables>
                 </configuration>
                 <dependencies>
                     <dependency>
                         <groupId>org.apache.maven.surefire</groupId>
                         <artifactId>surefire-junit47</artifactId>
-                        <version>3.0.0-M4</version>
+                        <version>3.0.0-M5</version>
                     </dependency>
                 </dependencies>
             </plugin>
