--- osgi-email-itest/src/test/java/hu/blackbelt/email/TestUtility.java
@@ -1,174 +1,0 @@
-package hu.blackbelt.email;
-
-import org.ops4j.pax.exam.TimeoutException;
-import org.osgi.framework.*;
-import org.osgi.util.tracker.ServiceTracker;
-
-import java.io.File;
-import java.io.IOException;
-import java.net.ConnectException;
-import java.net.HttpURLConnection;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.*;
-
-import static junit.framework.TestCase.assertNotNull;
-import static org.junit.Assert.assertEquals;
-
-public class TestUtility {
-
-    public static final Integer SERVICE_TIMEOUT = 30000;
-
-    private TestUtility() {
-        super();
-    }
-
-    /**
-     * Explodes the dictionary into a ,-delimited list of key=value pairs
-     */
-    public static String explode(Dictionary dictionary) {
-        Enumeration keys = dictionary.keys();
-        StringBuffer result = new StringBuffer();
-        while (keys.hasMoreElements()) {
-            Object key = keys.nextElement();
-            result.append(String.format("%s=%s", key, dictionary.get(key)));
-            if (keys.hasMoreElements()) {
-                result.append(", ");
-            }
-        }
-        return result.toString();
-    }
-
-    /*
-     * Provides an iterable collection of references, even if the original array
-     * is null
-     */
-    public static Collection<ServiceReference> asCollection(ServiceReference[] references) {
-        return references != null ? Arrays.asList(references) : Collections.<ServiceReference> emptyList();
-    }
-
-
-    public static  <T> T getOsgiService(BundleContext bundleContext, Class<T> type, long timeout) {
-        return getOsgiService(bundleContext, type, null, timeout);
-    }
-
-    public static  <T> T getOsgiService(BundleContext bundleContext, Class<T> type) {
-        return getOsgiService(bundleContext, type, null, SERVICE_TIMEOUT);
-    }
-
-    public static  <T> T getOsgiService(BundleContext bundleContext, Class<T> type, String filter, long timeout) {
-        ServiceTracker tracker = null;
-        try {
-            String flt;
-            if (filter != null) {
-                if (filter.startsWith("(")) {
-                    flt = "(&(" + Constants.OBJECTCLASS + "=" + type.getName() + ")" + filter + ")";
-                } else {
-                    flt = "(&(" + Constants.OBJECTCLASS + "=" + type.getName() + ")(" + filter + "))";
-                }
-            } else {
-                flt = "(" + Constants.OBJECTCLASS + "=" + type.getName() + ")";
-            }
-            Filter osgiFilter = FrameworkUtil.createFilter(flt);
-            tracker = new ServiceTracker(bundleContext, osgiFilter, null);
-            tracker.open(true);
-            // Note that the tracker is not closed to keep the reference
-            // This is buggy, as the service reference may change i think
-            Object svc = type.cast(tracker.waitForService(timeout));
-            if (svc == null) {
-                Dictionary dic = bundleContext.getBundle().getHeaders();
-                System.err.println("Test bundle headers: " + explode(dic));
-
-                for (ServiceReference ref : asCollection(bundleContext.getAllServiceReferences(null, null))) {
-                    System.err.println("ServiceReference: " + ref);
-                }
-
-                for (ServiceReference ref : asCollection(bundleContext.getAllServiceReferences(null, flt))) {
-                    System.err.println("Filtered ServiceReference: " + ref);
-                }
-
-                throw new RuntimeException("Gave up waiting for service " + flt);
-            }
-            return type.cast(svc);
-        } catch (InvalidSyntaxException e) {
-            throw new IllegalArgumentException("Invalid filter", e);
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-
-    public static File testTargetDir(Class clazz){
-        String relPath = clazz.getProtectionDomain().getCodeSource().getLocation().getFile();
-        File targetDir = new File(relPath);
-        if(!targetDir.exists()) {
-            targetDir.mkdir();
-        }
-        return targetDir;
-    }
-
-
-    public static void assertBundleStarted(BundleContext bundleContext, String name) {
-        Bundle bundle = findBundleByName(bundleContext, name);
-        assertNotNull("Bundle " + name + " should be installed", bundle);
-        assertEquals("Bundle " + name + " should be started", Bundle.ACTIVE, bundle.getState());
-    }
-
-    public static  Bundle findBundleByName(BundleContext bundleContext, String symbolicName) {
-        for (Bundle bundle : bundleContext.getBundles()) {
-            if (bundle.getSymbolicName().equals(symbolicName)) {
-                return bundle;
-            }
-        }
-        return null;
-    }
-
-    public static  void waitWebPage(String urlSt) throws InterruptedException, TimeoutException {
-        System.out.println("Waiting for url " + urlSt);
-        HttpURLConnection con = null;
-        long startTime = System.currentTimeMillis();
-        while (true) {
-            try {
-                URL url = new URL(urlSt);
-                con = (HttpURLConnection)url.openConnection();
-                int status = con.getResponseCode();
-                if (status == 200) {
-                    return;
-                }
-            } catch (ConnectException e) {
-                // Ignore connection refused
-            } catch (MalformedURLException e) {
-                throw new RuntimeException(e.getMessage(), e);
-            } catch (IOException e) {
-                throw new RuntimeException(e.getMessage(), e);
-            } finally {
-                if (con != null) {
-                    con.disconnect();
-                }
-            }
-            sleepOrTimeout(startTime, 20, "Timeout waiting for web page " + urlSt);
-        }
-    }
-
-    /**
-     * Sleeps for a short interval, throwing an exception if timeout has been reached. Used to facilitate a
-     * retry interval with timeout when used in a loop.
-     *
-     * @param startTime the start time of the entire operation in milliseconds
-     * @param timeout the timeout duration for the entire operation in seconds
-     * @param message the error message to use when timeout occurs
-     * @throws InterruptedException if interrupted while sleeping
-     */
-    public static void sleepOrTimeout(long startTime, long timeout, String message)
-            throws InterruptedException, TimeoutException {
-        timeout *= 1000; // seconds to millis
-        long elapsed = System.currentTimeMillis() - startTime;
-        long remaining = timeout - elapsed;
-        if (remaining <= 0) {
-            throw new TimeoutException(message);
-        }
-        long interval = Math.min(remaining, 1000);
-        Thread.sleep(interval);
-    }
-
-}
