+++ osgi-email-itest/src/test/resources/test-features.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<features name="judo-TEST"
+          xmlns="http://karaf.apache.org/xmlns/features/v1.5.0"
+          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+          xsi:schemaLocation="http://karaf.apache.org/xmlns/features/v1.5.0 http://karaf.apache.org/xmlns/features/v1.5.0">
+
+    <repository>mvn:hu.blackbelt.karaf.features/google-features/${google-karaf-features-version}/xml/features</repository>
+    <repository>mvn:hu.blackbelt.karaf.features/subethamail-features/${subethamail-karaf-features-version}/xml/features</repository>
+
+    <feature name="test"  version="1" description="Base test features" install="true">
+        <feature prerequisite="true">wrap</feature>
+        <feature prerequisite="true">scr</feature>
+        <feature>subethamail</feature>
+    </feature>
+</features>
