+++ osgi-email-karaf-features/pom.xml
@@ -119,7 +119,7 @@
                                         <descriptor>file:${project.build.directory}/feature/feature.xml</descriptor>
                                     </descriptors>
                                     <distribution>org.apache.karaf.features:framework</distribution>
-                                    <javase>1.8</javase>
+                                    <javase>11</javase>
                                     <framework>
                                         <feature>framework</feature>
                                         <feature>shell-compat</feature>
